Update directory processing service

+++ src/Training.FileExplorer.ServerApp/Training.FileExplorer.Application/Common/Querying/Extensions/LinqExtensions.cs
@@ -6,11 +6,17 @@ public static class LinqExtensions
 {
     public static IQueryable<TSource> ApplyPagination<TSource>(this IQueryable<TSource> source, FilterPagination paginationOptions)
     {
-        return source.Skip((paginationOptions.PageToken - 1) * paginationOptions.PageSize).Take(paginationOptions.PageSize);
+        // var pageSize = paginationOptions.DynamicPageSize;
+        // return source.Skip((int)((paginationOptions.PageToken - 1) * pageSize)).Take((int)pageSize);
+
+        return source.Skip((int)((paginationOptions.PageToken - 1) * paginationOptions.PageSize)).Take((int)paginationOptions.PageSize);
     }
 
     public static IEnumerable<TSource> ApplyPagination<TSource>(this IEnumerable<TSource> source, FilterPagination paginationOptions)
     {
-        return source.Skip((paginationOptions.PageToken - 1) * paginationOptions.PageSize).Take(paginationOptions.PageSize);
+        // var pageSize = paginationOptions.DynamicPageSize;
+        // return source.Skip((int)((paginationOptions.PageToken - 1) * pageSize)).Take((int)pageSize);
+
+        return source.Skip((int)((paginationOptions.PageToken - 1) * paginationOptions.PageSize)).Take((int)paginationOptions.PageSize);
     }
 }
